Share common components among hits components
