+++ site/gatsby-site/src/components/discover/hits/Details.js
@@ -1,20 +1,13 @@
 import React from 'react';
-import { Card, Button } from 'react-bootstrap';
+import { Card } from 'react-bootstrap';
 import { Highlight, Snippet } from 'react-instantsearch-dom';
 import { Image } from 'utils/cloudinary';
 import styled from 'styled-components';
 import { fill } from '@cloudinary/base/actions/resize';
-import WebArchiveLink from '../../WebArchiveLink';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faNewspaper,
-  faIdCard,
-  faUserShield,
-  faFlag,
-  faHashtag,
-} from '@fortawesome/free-solid-svg-icons';
+
 import md5 from 'md5';
 import { navigate } from 'gatsby';
+import Actions from './Actions';
 
 const IncidentCardImage = styled(Image)`
   height: ${({ height }) => height};
@@ -29,38 +22,6 @@ const StyledLabel = styled.p`
 const TitleHighlight = styled(Highlight)`
   font-weight: bold;
 `;
-
-const getFlagModalContent = () => (
-  <div className="modal-body">
-    <p>Is there a problem with this content? Examples of &quot;problems`&quot;` include,</p>
-    <ul>
-      <li>The text contents of the report are not parsed properly</li>
-      <li>The authors of the report are not associated with the report</li>
-      <li>The report is associated with the wrong incident</li>
-      <li>The text contents of the report are not parsed properly</li>
-    </ul>
-    <p>
-      Flagged content will still be displayed within the database, but a database editor will
-      periodically review the incident reports that have been flagged. Please note that the content
-      contained within incident reports (e.g., the commentary within a news article) does not need
-      to be correct or consistent across articles. If an article is wrong, misleading, or
-      fraudulent, the best response is to submit additional incident reports that correct the
-      record. The incident database is meant to capture the complete state of knowledge and
-      discourse for incidents, not as an arbiter of what happened in individual incidents. In future
-      versions of the database it will additionally be possible to apply tags to incident reports to
-      classNameify their content.
-    </p>
-    <p>Please do NOT flag content if,</p>
-    <ul>
-      <li>You disagree with the report</li>
-      <li>The linked report has disappeared from the web</li>
-      <li>The report should not be considered an `&quot;`incident`&quot;`</li>
-    </ul>
-    <button type="button" className="btn btn-danger btn-sm w-100" data-dismiss="modal">
-      Flag Report
-    </button>
-  </div>
-);
 
 export default function Details({
   item,
@@ -112,67 +73,13 @@ export default function Details({
       </Card.Body>
 
       <Card.Footer className="d-flex justify-content-between">
-        <WebArchiveLink
-          url={item.url}
-          date={item.date_submitted}
-          className="btn btn-link px-1"
-          title={'Authors'}
-        >
-          <FontAwesomeIcon icon={faNewspaper} className="fa-newspaper" title="Read the Source" />
-        </WebArchiveLink>
-
-        <Button
-          variant="link"
-          title="Authors"
-          onClick={() =>
-            authorsModal.openFor({
-              title: 'Authors',
-              body: () => item.authors.join(', '),
-            })
-          }
-        >
-          <FontAwesomeIcon icon={faIdCard} className="fa-id-card" />
-        </Button>
-
-        <Button
-          variant="link"
-          title="Submitters"
-          className="px-1"
-          onClick={() =>
-            submittersModal.openFor({
-              title: 'Submitters',
-              body: () => item.submitters.join(', '),
-            })
-          }
-        >
-          <FontAwesomeIcon icon={faUserShield} className="fa-user-shield" />
-        </Button>
-
-        <Button
-          variant="link"
-          title="Flag Report"
-          className="px-1"
-          onClick={() =>
-            flagReportModal.openFor({
-              title: 'Submitters',
-              body: getFlagModalContent,
-            })
-          }
-        >
-          <FontAwesomeIcon icon={faFlag} className="fa-flag" />
-        </Button>
-
-        {toggleFilterByIncidentId && (
-          <Button
-            variant="link"
-            aria-hidden="true"
-            className="d-flex align-items-center px-1"
-            onClick={() => toggleFilterByIncidentId(item.incident_id + '')}
-          >
-            <FontAwesomeIcon icon={faHashtag} className="fa-hashtag" title="Incident ID" />
-            {item.incident_id}
-          </Button>
-        )}
+        <Actions
+          authorsModal={authorsModal}
+          flagReportModal={flagReportModal}
+          submittersModal={submittersModal}
+          toggleFilterByIncidentId={toggleFilterByIncidentId}
+          item={item}
+        />
       </Card.Footer>
     </Card>
   );
